refactor(api): extract log-tail helpers from _log_stream_routes
_read_tail, _resolve_run_log, and _stream_terminal_state were raw
file I/O and log-parsing logic living directly in the Flask routes
module. Move them to a new sibling _log_tail_helpers.py, following
the split already established by _sse_log_helpers.py.

_resolve_run_log and _resolve_stream_log_path now take provider as
an explicit parameter instead of _resolve_run_log reaching for
Flask's current_app, keeping the new module Flask-independent like
its sibling.

Update tools/size_baseline.txt: register_log_stream_routes (still
grandfathered over the 50-line function cap) shifted from line 144
to line 54 now that the extracted functions are gone.
